Added 'Privacy' switch. Changed the styling.
Privacy switch will censor the text in the TextInput component when the
Switch is on.

Added the database breach occurance count to a descriptive sentence,
which will render when the data from the API call is finally processed.
The sentence will disappear when the search is initiated again.

Styling has been changed to provide a cleaner layout.