Added a dynamic "comment" <Text> and more styling.
Added "comment" <Text> component that changes it's value depending on
the value of this.state.hashes.  This was added to provide greater
clarity to the user as to whether or not they use should the password
they entered.

Changed the color of the background and the other componenets.  Also,
added some dynamic styling to the occurances <Text> component.