Added "Loading..." text, changed Button title to Search, added and comment to textResponseToJSON().
Changed the <Button> title from 'Go' to 'Search'.  It's a better description
of what the button does.

The user had no meaningful prompt as to what is going on when they tap the
"Search" button.  So, a new state proptery was added to indicate if the API
call and search of the results were on going.  If state.loading == true
then result will equal "Loading...".

Added a descriptive comment to textResponseToJSON().